requirements.txt
slither-analyzer==0.10.0
mythril==0.23.22
z3-solver
py-solc-x
requests
